Allow users to customize the style
Currently, the css style is copied directly from "style.css" after each
time Anki restarts. This makes any permanent changes to the Note Type
style impossible, unless the code is directly modified by the user.

The best solution is to separate the style in two. One for the addon
which lives in "style.css" and the other for users in "user_style.css".
The note type then imports both styles dynamically from the media folder.

This makes it possible for the addon to continue improving the style,
while also allowing users to customize it to their liking.