Reuse HTML template code for common note types
The front and back for the base note type shares the exact code, except
for the render() and show() function. Therefore, if makes more sense to
extract their common dependency into a template of its own, which makes
maintaining the code much easier.

In the shared render() and show() functions, the fields are read
dynamically as the card changes, toggling only the available fields.

This is the same for the cloze note type.