Display field title in preview and support multi-field notes
Currently, the Markdown preview only shows 2 Fields even if a note type
has more than 2, which can result in a messy output.

These changes make it possible for the Mardown preview to display each
field seperately and use its title instead of a generic name.

Unfortunately, there is some redundancy in the code and I don't know how
to reduce it yet, but we can just refactor it later.